refactor: grouped anime related files
