+++ src/controllers/anime/AnimeSearchController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
-import { RequestStatuses } from "../ts/enums";
-import AnimeFilterService from "../services/filters/AnimeFilterService";
+import { RequestStatuses } from "../../ts/enums";
+import AnimeFilterService from "../../services/filters/AnimeSearchFilter";
 
 export default class SearchController {
     public static async getAnime(req: Request, res: Response) {
