+++ src/index.ts
@@ -8,10 +8,10 @@ import { shikimoriRouter } from './routes/ShikimoriRouter';
 import { tokenRouter } from './routes/TokenRouter';
 import { watchListRouter } from './routes/WatchListRouter';
 import { followRouter } from './routes/FollowRouter';
-import { animeRouter } from './routes/AnimeRouter';
+import { animeRouter } from './routes/anime/AnimeRouter';
+import { searchRouter } from './routes/anime/AnimeSearchRouter';
 import { groupListRouter } from './routes/GroupListRouter';
 import { notificationRouter as notificationRouter } from './routes/NotificationRouter';
-import { searchRouter } from './routes/SearchRouter';
 import * as fs from 'fs';
 import * as https from 'https';
 import * as http from 'http';
