+++ src/requests/AnimeSearchRequest.ts
@@ -2,22 +2,20 @@ import { RequestAuthTypes } from "../ts/enums";
 import Request from "./Request";
 import {
     genresValidator,
-    yearsValidator,
     seasonsValidator,
     nameValidator,
-    episodeValidator,
     statusesValidator,
     rpaRatingsValidator,
     mediaTypesValidator
-} from "../validators/FilterValidator";
-import { softPeriodValidator } from '../validators/PeriodValidator';
+} from "../validators/AnimeValidator";
+import { softPeriodValidator } from "../validators/PeriodValidator";
 
 export default class SearchAnimeRequest extends Request {
 
     /**
      * Define auth type for this request
      */
-    protected authType = RequestAuthTypes.None;
+    protected authType = RequestAuthTypes.Optional;
 
     /**
      * define validation rules for this request
@@ -28,14 +26,12 @@ export default class SearchAnimeRequest extends Request {
             ...
             genresValidator("includeGenres"),
             genresValidator("excludeGenres"),
-            yearsValidator(),
             softPeriodValidator("period"),
-            seasonsValidator(),
-            nameValidator(),
-            episodeValidator(),
-            statusesValidator(),
-            rpaRatingsValidator(),
-            mediaTypesValidator()
+            seasonsValidator("seasons"),
+            nameValidator("name"),
+            statusesValidator("statuses"),
+            rpaRatingsValidator("rpaRatings"),
+            mediaTypesValidator("mediaTypes"),
         ]
     }
 }
