+++ src/services/anime/AnimeUpdateService.ts
@@ -1,11 +1,11 @@
 import { Anime, Prisma, User } from "@prisma/client";
-import ShikimoriApiService from "./shikimori/ShikimoriApiService";
-import { ServerError, ShikimoriAnimeFull, ShikimoriAnime } from "../ts/index";
-import { KodikAnimeFull, checkAnime, translation } from "../ts/kodik";
-import prisma from "../db";
-import { RequestStatuses } from "../ts/enums";
-import InternalServerError from "../errors/servererrors/InternalServerError";
-import { cyrillicSlug } from "../helper/cyrillic-slug";
+import ShikimoriApiService from "../shikimori/ShikimoriApiService";
+import { ServerError, ShikimoriAnimeFull, ShikimoriAnime } from "../../ts/index";
+import { KodikAnimeFull, checkAnime, translation } from "../../ts/kodik";
+import prisma from "../../db";
+import { RequestStatuses } from "../../ts/enums";
+import InternalServerError from "../../errors/servererrors/InternalServerError";
+import { cyrillicSlug } from "../../helper/cyrillic-slug";
 
 interface iAnimeUpdateService {
     shikimoriApi: ShikimoriApiService | undefined
